Fix C4242 warnings in liblzma: add explicit casts for narrowing conversions

+++ src/native/external/xz/src/liblzma/lzma/lzma_decoder.c
@@ -378,7 +378,7 @@ lzma_decode(void *coder_ptr, lzma_dict *restrict dictptr,
 			}
 
 			// Write decoded literal to dictionary
-			dict_put(&dict, symbol);
+			dict_put(&dict, (uint8_t)symbol);
 			continue;
 		}
 
@@ -742,7 +742,7 @@ slow:
 			}
 
 	case SEQ_LITERAL_WRITE:
-			if (dict_put_safe(&dict, symbol)) {
+			if (dict_put_safe(&dict, (uint8_t)symbol)) {
 				coder->sequence = SEQ_LITERAL_WRITE;
 				goto out;
 			}
